requirements.txt
Python: 3.10.12
PyTorch: 2.0.0
Torchvision: 0.15.1
CUDA: 11.8
CUDNN: 8900
NumPy: 1.24.3
PIL: 9.5.0